feat: Consolidate home module into feature
