+++ feature/home/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.hyunjung.cherrydan.android.feature.ui)
+}
+
+android {
+    namespace = "com.hyunjung.feature.home"
+}
+
+dependencies {
+    implementation(libs.androidx.navigation.compose)
+}
