+++ feature/home/src/main/java/com/hyunjung/feature/home/category/CategoryScreen.kt
@@ -1,0 +1,16 @@
+package com.hyunjung.feature.home.category
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.hyunjung.core.presentation.designsystem.CherrydanColors
+import com.hyunjung.core.presentation.designsystem.CherrydanTypography
+
+@Composable
+fun CategoryScreen(modifier: Modifier = Modifier) {
+    Text(
+        text = "카테고리",
+        style = CherrydanTypography.Title1,
+        color = CherrydanColors.Black
+    )
+}
