--- home/domain/src/main/java/com/hyunjung/home/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.hyunjung.home.domain
-
-class MyClass {
-}
