--- home/presentation/src/main/java/com/hyunjung/home/presentation/home/CategoryTab.kt
@@ -1,10 +1,0 @@
-package com.hyunjung.home.presentation.home
-
-enum class CategoryTab(val label: String) {
-    All("전체"),
-    Region("지역"),
-    Product("제품"),
-    Reporter("기자단"),
-    SNS("SNS 플랫폼"),
-    Experience("체험단 플랫폼")
-}
