+++ settings.gradle.kts
@@ -30,9 +30,6 @@ include(":core:domain")
 include(":core:data")
 include(":core:database")
 include(":core:network")
-include(":home:data")
-include(":home:domain")
-include(":home:presentation")
 include(":notification:data")
 include(":notification:domain")
 include(":notification:presentation")
@@ -41,3 +38,4 @@ include(":search:domain")
 include(":search:data")
 include(":core:model")
 include(":feature:auth")
+include(":feature:home")
